Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.0
opencv-python==4.12.0.88
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0
transformers==4.57.3
torch==2.9.0
tensorflow==2.19.0
peft==0.18.0
accelerate==1.1.1
sentencepiece==0.2.0
numpy>=2.0.0,<2.3.0
textblob==0.19.0
safetensors==0.4.5
huggingface-hub>=0.34.0,<1.0.0

azure-ai-formrecognizer 
pillow
azure-core
google-generativeai>=0.7.2
pymongo>=4.6
dnspython>=2.0
certifi>=2024.0.0

paddleocr>=2.7.0
paddlepaddle>=2.6.0
datasets>=2.20.0
layoutparser>=0.3.4
pandas>=2.2.0
openpyxl>=3.1.0
